need to fix orientation

+++ colorlabeler.py
@@ -25,4 +25,4 @@ def findTankCentroid(img, color = "green"):
 				num_hits += 1
 	mask[int(x / num_hits)][int(y / num_hits)] = 150
 	cv2.imwrite("yo.png", mask)
-	return (x / num_hits, y / num_hits)
+	return (int(x / num_hits), int(y / num_hits))
